requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
tqdm==4.66.5
pymupdf==1.24.10
chromadb==0.5.5
ollama==0.3.3
